bles/requirements.txt
Django==1.8.4
Pillow==2.9.0
django-extensions==1.5.5
django-push-notifications==1.3.1
djangorestframework==3.2.2
six==1.9.0
django-datatables-view==1.12.1
